service selector for daemonset only
Since daemonset and statefulset templates have the same labels, the
service selector was selecting the statefulset pods as well. Due to
this, some of the requests were being sent to the statefulset pods,
resulting in `Connection refused`, because those pods don't expose any
port.

This change adds a new label `kind` whose value is the kind of the
resource group and sets service selector to select the daemonset kind.